requirement.txt
amqp==2.2.1
anyjson==0.3.3
apipkg==1.4
appdirs==1.4.3
APScheduler==3.3.1
##apt-xapian-index==0.45
attrs==17.3.0
Automat==0.6.0
Babel==2.5.1
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
billiard==3.5.0.3
boto3==1.5.20
botocore==1.8.50
cachetools==2.0.1
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
ChatterBot==0.6.2
chatterbot-corpus==1.0.1
Cheetah==2.4.4
cl==0.0.3
click==6.7
cloudpickle==0.5.2
colorama==0.2.5
coloredlogs==9.0
configobj==4.7.2
configparser==3.5.0
constantly==15.1.0
coverage==4.0.3
crypto==1.4.1
cycler==0.10.0
cymem==1.31.2
Cython==0.27.3
cytoolz==0.8.2
dill==0.2.7.1
##distro-info==0.12
Django==1.11
djangorestframework==3.7.7
dnspython==1.15.0
docopt==0.6.2
docutils==0.14
duckling==1.7.1
email-validator==1.0.3
emoji==0.4.5
##en-core-web-sm==1.2.0
enum==0.4.6
enum34==1.1.6
et-xmlfile==1.0.1
execnet==1.4.1
fake-useragent==0.1.10
Flask==0.12.2
Flask-Cors==3.0.3
flower==0.9.2
fs==2.0.20
ftfy==4.4.3
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==2.2.0
fuzzywuzzy==0.15.1
geopy==1.11.0
gevent==1.2.2
google-auth==1.2.1
google-auth-httplib2==0.0.3
google-cloud-core==0.24.1
google-cloud-storage==1.0.0
googleapis-common-protos==1.5.3
greenlet==0.4.12
##gyp==0.1
headers-workaround==0.18
html5lib==1.0b8
httplib2==0.10.3
humanfriendly==4.10
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
instagram-scraper==1.5.32
instalooter==1.0.0a5
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
jmespath==0.9.3
JPype1==0.6.2
jsondatabase==0.1.7
jsonpatch==1.3
jsonpointer==1.0
jsonschema==2.6.0
klein==17.10.0
kombu==4.1.0
##Landscape-Client==14.12
librabbitmq==1.0.3
line-bot-sdk==1.5.0
lml==0.0.1
lxml==4.1.1
mailer==0.7
MarkupSafe==1.0
mathparse==0.1.1
matplotlib==2.1.0
meld3==0.6.10
##mitie==0.2.0
mock==2.0.0
monotonic==1.4
msgpack-numpy==0.4.1
msgpack-python==0.5.4
murmurhash==0.28.0
mysql==0.0.1
MySQL-python==1.2.5
Naked==0.1.31
nltk==3.2.5
numpy==1.14.2
oauth==1.0.1
oauthlib==2.0.6
olefile==0.44
openpyxl==2.4.9
##PAM==0.4.2
pathlib==1.0.1
pbr==3.1.1
pdfkit==0.6.1
pep8==1.7.0
Pillow==4.2.1
plac==0.9.6
preshed==1.0.0
prettytable==0.7.2
protobuf==3.5.0.post1
psutil==5.2.2
py==1.4.34
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pycurl==7.19.3
pyexcel-io==0.5.4
pyexcel-xlsx==0.5.4
PyJWT==1.5.3
pyliblzma==0.5.3
pymessenger==0.0.7.0
pymongo==3.6.0
pyOpenSSL==17.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
pyserial==2.6
pytest==3.0.7
pytest-cache==1.0
pytest-cov==2.5.1
pytest-flask==0.10.0
pytest-pep8==1.0.6
pytest-services==1.1.14
##python-apt===0.9.3.5ubuntu2
python-coveralls==2.9.1
python-crfsuite==0.9.2
python-dateutil==2.6.1
##python-debian===0.1.21-nmu2ubuntu2
python-Levenshtein==0.12.0
python-magic==0.4.13
python-memcached==1.53
python-openid==2.2.5
python-social-auth==0.3.6
python-twitter==3.3
pytz==2017.3
PyYAML==3.12
##qiscusapi==0.0.8
##-e git+https://github.com/RasaHQ/rasa_nlu.git@308a25061dc249113993542e2cc14c766a0c2caa#egg=rasa_nlu
redis==2.10.6
regex==2017.4.5
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rest-social-auth==1.2.0
rivescript==1.14.7
rsa==3.4.2
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.0
s3transfer==0.1.12
scikit-learn==0.18.1
scipy==0.19.0
service-identity==17.0.0
shellescape==3.4.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
sklearn-crfsuite==0.3.5
social-auth-app-django==1.2.0
social-auth-core==1.5.0
SQLAlchemy==1.1.15
##sqlparser==1.0
stevedore==1.28.0
subprocess32==3.2.7
supervisor==3.0b2
tabulate==0.7.7
tenacity==4.9.0
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
tornado==4.5.2
tqdm==4.23.0
treq==17.3.1
Twisted==17.9.0
##Twisted-Core==13.2.0
##Twisted-Names==13.2.0
##Twisted-Web==13.2.0
typing==3.6.2
tzlocal==1.4
uber-rides==0.4.1
ujson==1.35
urllib3==1.22
vine==1.1.4
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
##webkit2png==0.8.3
Werkzeug==0.13
wrapt==1.10.11
##XlsxWriter==1.0.2
##zope.interface==4.4.3
